+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.21.2 - Unreleased
 ### Fixed
 - Fixed a bug in which the times in the x-axis of the reports chart was not being converted to the local timezone ([#274](https://github.com/putyourlightson/craft-campaign/issues/274)).
+- Fixed a bug in which the `getMailingListBySlug` method did not specify the current site ([#278](https://github.com/putyourlightson/craft-campaign/issues/278)).
 
 ## 1.21.1 - 2021-10-07
 ### Changed